Cleaned up code, moved pages out of components, fixed most react warnings in console
