+++ client/public/index.html
@@ -1,37 +1,38 @@
-    <!DOCTYPE html>
-    <html lang="en">
+<!DOCTYPE html>
+<html lang="en">
 
-    <head>
-        <title>Code-Crush</title>
-        <meta charset="utf-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-        <meta name="description" content="codecrush">
-        <meta name="author" content="Devcrud">
-        <!-- Bootstrap + codecrush main styles -->
-        <link rel="stylesheet" href="assets/css/codecrush.css">
-        <!-- font icons -->
-        <link rel="stylesheet" href="assets/vendors/themify-icons/css/themify-icons.css">
-        <link rel="stylesheet" href="assets/vendors/animate/animate.css">
+<head>
+  <title>Code-Crush</title>
+  <meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+  <meta name="description" content="codecrush">
+  <meta name="author" content="Devcrud">
+  <!-- Bootstrap + codecrush main styles -->
+  <link rel="stylesheet" href="assets/css/codecrush.css">
+  <!-- font icons -->
+  <link rel="stylesheet" href="assets/vendors/themify-icons/css/themify-icons.css">
+  <link rel="stylesheet" href="assets/vendors/animate/animate.css">
 
-    </head>
+</head>
 
 
-    <body data-spy="scroll" data-target=".navbar" data-offset="40" id="home">
-      <noscript>You need to enable JavaScript to run this app.</noscript>
-      <div id="root"></div>  
+<body data-spy="scroll" data-target=".navbar" data-offset="40" id="home">
+  <noscript>You need to enable JavaScript to run this app.</noscript>
+  <div id="root"></div>
 
-      <!-- core  -->
-      <script src="assets/vendors/jquery/jquery-3.4.1.js"></script>
-      <script src="assets/vendors/bootstrap/bootstrap.bundle.js"></script>
+  <!-- core  -->
+  <script src="assets/vendors/jquery/jquery-3.4.1.js"></script>
+  <script src="assets/vendors/bootstrap/bootstrap.bundle.js"></script>
 
-      <!-- bootstrap affix --> 
-      <script src="assets/vendors/bootstrap/bootstrap.affix.js"></script>
+  <!-- bootstrap affix -->
+  <script src="assets/vendors/bootstrap/bootstrap.affix.js"></script>
 
-       <!-- wow.js --> 
-      
+  <!-- wow.js -->
 
-       <!-- codecrush js --> 
-      <script src="assets/js/codecrush.js"></script>
 
-  </body>
-  </html>
+  <!-- codecrush js -->
+  <script src="assets/js/codecrush.js"></script>
+
+</body>
+
+</html>
