+++ client/src/components/Head.js
@@ -1,9 +1,8 @@
-import * as React from "react";
+import React from "react";
 
 export default function Head() {
-    return (
+  return (
     <div>
-        
     </div>
-    );
-  }
+  );
+}
