+++ client/src/components/images.js
@@ -1,16 +1,16 @@
 const IMAGES = {
-    aboutSection: require("../assets/imgs/about-section.jpg"),
-    ConnectWithPeopleEverywhere: require("../assets/imgs/ConnectWithPeopleEverywhere.jpg"),
-    ExploreYourPassions: require("../assets/imgs/ExploreYourPassions.jpg"),
-    FacetoFace: require("../assets/imgs/FacetoFace.jpg"),
-    finger: require("../assets/imgs/finger.jpg"),
-    Inclusive: require("../assets/imgs/Inclusive.jpg"),
-    logo: require("../assets/imgs/logo.png"),
-    main: require("../assets/imgs/main.jpg"),
-    RollWithUsAnywhere: require("../assets/imgs/RollWithUsAnywhere.jpg"),
-    verify: require("../assets/imgs/verify.jpg"),
-    who: require("../assets/imgs/who.jpg"),
-    // : require("../assets/imgs/.jpg"),
+  aboutSection: require("../assets/imgs/about-section.jpg"),
+  ConnectWithPeopleEverywhere: require("../assets/imgs/ConnectWithPeopleEverywhere.jpg"),
+  ExploreYourPassions: require("../assets/imgs/ExploreYourPassions.jpg"),
+  FacetoFace: require("../assets/imgs/FacetoFace.jpg"),
+  finger: require("../assets/imgs/finger.jpg"),
+  Inclusive: require("../assets/imgs/Inclusive.jpg"),
+  logo: require("../assets/imgs/logo.png"),
+  main: require("../assets/imgs/main.jpg"),
+  RollWithUsAnywhere: require("../assets/imgs/RollWithUsAnywhere.jpg"),
+  verify: require("../assets/imgs/verify.jpg"),
+  who: require("../assets/imgs/who.jpg"),
+  // : require("../assets/imgs/.jpg"),
 }
 
 export default IMAGES;
