+++ client/src/index.js
@@ -2,5 +2,4 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
 
-
 ReactDOM.render(<App />, document.getElementById('root'));
